+++ PCSX2_Core/Source/Resources/EE/INTC/Types/EEIntcRegisters_t.h
@@ -14,21 +14,21 @@ class EEIntcRegister_MASK_t : public BitfieldRegister32_t
 public:
 	struct Fields
 	{
-		static constexpr u8 GS = 0;
-		static constexpr u8 SBUS = 1;
-		static constexpr u8 VBON = 2;
-		static constexpr u8 VBOF = 3;
-		static constexpr u8 VIF0 = 4;
-		static constexpr u8 VIF1 = 5;
-		static constexpr u8 VU0 = 6;
-		static constexpr u8 VU1 = 7;
-		static constexpr u8 IPU = 8;
-		static constexpr u8 TIM0 = 9;
-		static constexpr u8 TIM1 = 10;
-		static constexpr u8 TIM2 = 11;
-		static constexpr u8 TIM3 = 12;
-		static constexpr u8 SFIFO = 13;
-		static constexpr u8 VU0WD = 14;
+		static constexpr int GS = 0;
+		static constexpr int SBUS = 1;
+		static constexpr int VBON = 2;
+		static constexpr int VBOF = 3;
+		static constexpr int VIF0 = 4;
+		static constexpr int VIF1 = 5;
+		static constexpr int VU0 = 6;
+		static constexpr int VU1 = 7;
+		static constexpr int IPU = 8;
+		static constexpr int TIM0 = 9;
+		static constexpr int TIM1 = 10;
+		static constexpr int TIM2 = 11;
+		static constexpr int TIM3 = 12;
+		static constexpr int SFIFO = 13;
+		static constexpr int VU0WD = 14;
 	};
 
 	EEIntcRegister_MASK_t();
@@ -48,26 +48,26 @@ class EEIntcRegister_STAT_t : public BitfieldRegister32_t
 public:
 	struct Fields
 	{
-		static constexpr u8 GS = 0;
-		static constexpr u8 SBUS = 1;
-		static constexpr u8 VBON = 2;
-		static constexpr u8 VBOF = 3;
-		static constexpr u8 VIF0 = 4;
-		static constexpr u8 VIF1 = 5;
-		static constexpr u8 VU0 = 6;
-		static constexpr u8 VU1 = 7;
-		static constexpr u8 IPU = 8;
-		static constexpr u8 TIM0 = 9;
-		static constexpr u8 TIM1 = 10;
-		static constexpr u8 TIM2 = 11;
-		static constexpr u8 TIM3 = 12;
-		static constexpr u8 SFIFO = 13;
-		static constexpr u8 VU0WD = 14;
+		static constexpr int GS = 0;
+		static constexpr int SBUS = 1;
+		static constexpr int VBON = 2;
+		static constexpr int VBOF = 3;
+		static constexpr int VIF0 = 4;
+		static constexpr int VIF1 = 5;
+		static constexpr int VU0 = 6;
+		static constexpr int VU1 = 7;
+		static constexpr int IPU = 8;
+		static constexpr int TIM0 = 9;
+		static constexpr int TIM1 = 10;
+		static constexpr int TIM2 = 11;
+		static constexpr int TIM3 = 12;
+		static constexpr int SFIFO = 13;
+		static constexpr int VU0WD = 14;
 
-		static constexpr u8 IRQ_KEYS[Constants::EE::INTC::NUMBER_IRQ_LINES] = { GS, SBUS, VBON, VBOF, VIF0, VIF1, VU0, VU1, IPU, TIM0, TIM1, TIM2, TIM3, SFIFO, VU0WD };
-		static constexpr u8 VIF_KEYS[Constants::EE::VPU::VIF::NUMBER_VIF_CORES] = { VIF0, VIF1 };
-		static constexpr u8 VU_KEYS[Constants::EE::VPU::VU::NUMBER_VU_CORES] = { VU0, VU1 };
-		static constexpr u8 TIM_KEYS[Constants::EE::Timers::NUMBER_TIMERS] = { TIM0, TIM1, TIM2, TIM3 };
+		static constexpr int IRQ_KEYS[Constants::EE::INTC::NUMBER_IRQ_LINES] = { GS, SBUS, VBON, VBOF, VIF0, VIF1, VU0, VU1, IPU, TIM0, TIM1, TIM2, TIM3, SFIFO, VU0WD };
+		static constexpr int VIF_KEYS[Constants::EE::VPU::VIF::NUMBER_VIF_CORES] = { VIF0, VIF1 };
+		static constexpr int VU_KEYS[Constants::EE::VPU::VU::NUMBER_VU_CORES] = { VU0, VU1 };
+		static constexpr int TIM_KEYS[Constants::EE::Timers::NUMBER_TIMERS] = { TIM0, TIM1, TIM2, TIM3 };
 	};
 
 	explicit EEIntcRegister_STAT_t(const std::shared_ptr<EEIntcRegister_MASK_t> & mask);
